+++ doc/changelog.rst
@@ -9,6 +9,10 @@ Added
 - Support for Python 3.14.
 - Compile regexes.
 
+Fixed
+^^^^^
+- Sub-attributes of requested complex attributes are now included in responses. :issue:`114`
+
 Removed
 ^^^^^^^
 - Support for Python 3.9.